Paddle class moved to movable package
